+++ delivery/src/main/java/com/delivery/delivery/dto/auth/ProdutoResponse.java
@@ -1,6 +1,8 @@
 package com.delivery.delivery.dto.auth;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -8,7 +10,7 @@ public class ProdutoResponse {
     private Integer id;
     private String nmProduto;
     private String dsProduto;
-    private Double vlPreco;
+    private BigDecimal vlPreco;
     private Boolean flAtivo;
     private LocalDateTime dtCriacao;
     private LocalDateTime dtAtualizacao;
